Add jobsStatistics GraphQL Query
I did not use/change the existing functions in the
JobRepository because they did not match the GraphQL API and
are used in the JobCache.

Doctrin's QueryBuilder did not let me use integer division or
the floor function, so the histWalltime query is basically broken
and takes up >1G memory. I worked around that, but there must be
a better solution.